Remove the buildscript bloat
