--- buildSrc/src/main/kotlin/fixed-minecraft.compilation.gradle.kts
@@ -1,19 +1,0 @@
-import org.jetbrains.kotlin.gradle.dsl.KotlinVersion
-
-plugins {
-    java
-    kotlin("jvm")
-}
-
-java {
-    withSourcesJar()
-}
-
-kotlin {
-    jvmToolchain(17)
-
-    compilerOptions {
-        apiVersion = KotlinVersion.KOTLIN_1_9
-        languageVersion = KotlinVersion.KOTLIN_1_9
-    }
-}
