+++ src/main/kotlin/net/greenjab/fixedminecraft/FixedMinecraftDataGenerator.kt
@@ -5,6 +5,6 @@ import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator
 
 object FixedMinecraftDataGenerator : DataGeneratorEntrypoint {
     override fun onInitializeDataGenerator(fabricDataGenerator: FabricDataGenerator) {
-        TODO("Not yet implemented")
+
     }
 }
